#388 create AcctBridge class and add to configurations
