+++ src/f09_brick/brick_config.py
@@ -208,6 +208,14 @@ def brick_format_00041_bridge_nub_label_v0_0_0() -> str:
     return "brick_format_00041_bridge_nub_label_v0_0_0"
 
 
+def brick_format_00042_bridge_group_id_v0_0_0() -> str:
+    return "brick_format_00042_bridge_group_id_v0_0_0"
+
+
+def brick_format_00043_bridge_acct_id_v0_0_0() -> str:
+    return "brick_format_00043_bridge_acct_id_v0_0_0"
+
+
 def brick_format_00113_acct_otx2inx_v0_0_0() -> str:
     return "brick_format_00113_acct_otx2inx_v0_0_0"
 
@@ -241,6 +249,8 @@ def get_brick_format_filenames() -> set[str]:
         brick_format_00036_problem_healer_v0_0_0(),
         brick_format_00040_bridge_otx2inx_v0_0_0(),
         brick_format_00041_bridge_nub_label_v0_0_0(),
+        brick_format_00042_bridge_group_id_v0_0_0(),
+        brick_format_00043_bridge_acct_id_v0_0_0(),
         brick_format_00113_acct_otx2inx_v0_0_0(),
         brick_format_00114_acct_nub_label_v0_0_0(),
     }
@@ -271,6 +281,8 @@ def get_brick_numbers() -> set[str]:
         "br00036",
         "br00040",
         "br00041",
+        "br00042",
+        "br00043",
         "br00113",
         "br00114",
     }
@@ -308,6 +320,8 @@ def get_brick_format_headers() -> dict[str, list[str]]:
         "fiscal_id,healer_id,label,owner_id,parent_road,problem_bool": brick_format_00036_problem_healer_v0_0_0(),
         "inx_wall,inx_word,jaar_type,otx_wall,otx_word,unknown_word": brick_format_00040_bridge_otx2inx_v0_0_0(),
         "inx_label,inx_wall,otx_label,otx_wall,unknown_word": brick_format_00041_bridge_nub_label_v0_0_0(),
+        "inx_group_id,inx_wall,otx_group_id,otx_wall,unknown_word": brick_format_00042_bridge_group_id_v0_0_0(),
+        "inx_acct_id,inx_wall,otx_acct_id,otx_wall,unknown_word": brick_format_00043_bridge_acct_id_v0_0_0(),
         "acct_id,fiscal_id,inx_word,jaar_type,otx_word,owner_id": brick_format_00113_acct_otx2inx_v0_0_0(),
         "acct_id,fiscal_id,inx_label,jaar_type,otx_label,owner_id": brick_format_00114_acct_nub_label_v0_0_0(),
     }
@@ -387,6 +401,10 @@ def get_brick_elements_sort_order() -> list[str]:
         "cumlative_day",
         "weekday_label",
         "weekday_order",
+        "otx_acct_id",
+        "inx_acct_id",
+        "otx_group_id",
+        "inx_group_id",
         "otx_word",
         "inx_word",
         "otx_label",
@@ -465,6 +483,10 @@ def get_brick_sqlite_type() -> dict[str, str]:
         "otx_wall": "TEXT",
         "inx_wall": "TEXT",
         "unknown_word": "TEXT",
+        "otx_acct_id": "TEXT",
+        "inx_acct_id": "TEXT",
+        "otx_group_id": "TEXT",
+        "inx_group_id": "TEXT",
         "otx_word": "TEXT",
         "inx_word": "TEXT",
         "otx_label": "TEXT",
@@ -496,6 +518,8 @@ def get_brick_category_ref() -> dict[str, set[str]]:
         "bud_item_reason_premiseunit": ["br00026"],
         "bud_item_reasonunit": ["br00027"],
         "budunit": ["br00029"],
+        "bridge_group_id": ["br00042"],
+        "bridge_acct_id": ["br00043"],
         "bridge_otx2inx": ["br00040", "br00113"],
         "bridge_nub_label": ["br00041", "br00114"],
     }
