+++ src/f09_brick/brick_formats/brick_format_00043_bridge_acct_id_v0_0_0.json
@@ -1,0 +1,29 @@
+{
+    "attributes": {
+        "event_id": {
+            "otx_key": true
+        },
+        "face_id": {
+            "otx_key": true
+        },
+        "inx_wall": {
+            "otx_key": false
+        },
+        "inx_acct_id": {
+            "otx_key": false
+        },
+        "otx_wall": {
+            "otx_key": false
+        },
+        "otx_acct_id": {
+            "otx_key": true
+        },
+        "unknown_word": {
+            "otx_key": false
+        }
+    },
+    "brick_number": "br00043",
+    "categorys": [
+        "bridge_acct_id"
+    ]
+}
